[Lang] Move optional-annotation parsing into its own module
The `_split_optional` / `_is_optional_capable` helpers are self-contained (they touch only external imports and are
called from one site) yet lived in the large, central `_func_base.py`. Move them to a new
`lang/_optional_annotation.py` (exposing `split_optional`) to keep the feature partitioned from the kernel base
class, per AGENTS.md. No behaviour change.
